backend/requirements.txt
fastapi~=0.88.0
SQLAlchemy~=1.4.44
pydantic~=1.10.2
psycopg2-binary~=2.9.5
uvicorn~=0.20.0
python-dotenv~=0.21.0
passlib~=1.7.4
python-multipart~=0.0.5
python-jose~=3.3.0
